优化功能

+++ src/Commands/BuildCreateCommand.php
@@ -15,7 +15,7 @@ class BuildCreateCommand extends Command
      *
      * @var string
      */
-    protected $signature = 'hd:build {model} {module}';
+    protected $signature = 'hd:build {model} {module} {model_title} {module_title}';
 
     /**
      * The console command description.
@@ -26,6 +26,8 @@ class BuildCreateCommand extends Command
     protected $module;
     protected $model;
     protected $modelInstance;
+    protected $modelTitle;
+    protected $moduleTitle;
 
     /**
      * Create a new command instance.
@@ -41,8 +43,10 @@ class BuildCreateCommand extends Command
     {
         $this->model  = ucfirst($this->argument('model'));
         $this->module = ucfirst($this->argument('module'));
+        $this->modelTitle = ucfirst($this->argument('model_title'));
+        $this->moduleTitle = ucfirst($this->argument('module_title'));
         $this->setVars($this->model, $this->module);
-        //if ( ! $this->check()) {
+        if ($this->check()) {
             $this->setModelInstance();
             $this->setVar('MODEL_TITLE', $this->getTableComment($this->modelInstance));
             $this->call('hd:handle', ['model' => $this->model, 'module' => $this->module]);
@@ -50,7 +54,37 @@ class BuildCreateCommand extends Command
             $this->createRequest();
             $this->createRoute();
             $this->createViews();
-        //}
+            $this->setModelFillable();
+            $this->setModuleMenus();
+        }
+    }
+
+    protected function setModuleMenus()
+    {
+        $file =$this->getVar('MODULE_PATH').'config/menus.php';
+        $menus = include $file;
+        if ( ! isset($menus[$this->getVar('SMODULE')])) {
+            $menus[$this->getVar('SMODULE')] = [
+                "title"      => "{$this->moduleTitle}管理",
+                "icon"       => "fa fa-navicon",
+                'permission' => '权限标识',
+                "menus"      => []
+            ];
+        }
+        $menus[$this->getVar('SMODULE')]['menus'][]=
+            ["title" => "{$this->modelTitle}管理", "permission" => '', "url" => "/{$this->vars['SMODULE']}/{$this->vars['SMODEL']}"]
+        ;
+        file_put_contents($file,'<?php return '.var_export($menus,true).';');
+    }
+
+    protected function setModelFillable()
+    {
+        $columns = array_keys($this->getColumnData($this->modelInstance));
+        $columns = "['".implode("','", $columns)."'];";
+        $model   = $this->vars['MODEL_PATH'].$this->model.'.php';
+        $content = file_get_contents($model);
+        $content = preg_replace('@(protected\s+\$fillable\s*=\s*)(.*?);@', '${1}'.$columns, $content);
+        file_put_contents($model, $content);
     }
 
     protected function check(): bool
@@ -58,12 +92,6 @@ class BuildCreateCommand extends Command
         $model = $this->vars['NAMESPACE'].'Entities\\'.$this->model;
         if ( ! class_exists($model)) {
             $this->error("module {$model} no exists");
-
-            return false;
-        }
-        $file = $this->getVar('CONTROLLER_PATH').$this->model.'Controller.php';
-        if (is_file($file)) {
-            $this->info('buile file is exists');
 
             return false;
         }
@@ -78,7 +106,10 @@ class BuildCreateCommand extends Command
             $dir     = $this->vars['MODULE_PATH']."Resources/views/{$this->vars['SMODEL']}/";
             is_dir($dir) or mkdir($dir, 0755, true);
             Storage::makeDirectory($dir);
-            file_put_contents($dir."{$name}.blade.php", $content);
+            $file = $dir."{$name}.blade.php";
+            if ( ! is_file($file)) {
+                file_put_contents($file, $content);
+            }
         }
         $this->info('view create successflly');
     }
@@ -111,18 +142,28 @@ str;
 
     public function createController()
     {
+        $file = $this->getVar('CONTROLLER_PATH').$this->model.'Controller.php';
+        if (is_file($file)) {
+            $this->info('controller file is exists');
+
+            return false;
+        }
         $content = $this->replaceVars(__DIR__.'/build/controller.tpl');
-        $file    = $this->getVar('CONTROLLER_PATH').$this->model.'Controller.php';
         file_put_contents($file, $content);
         $this->info('controller create successflly');
     }
 
     public function createRequest()
     {
+        $file = $this->getVar('REQUEST_PATH').$this->model.'Request.php';
+        if (is_file($file)) {
+            $this->info('request file is exists');
+
+            return false;
+        }
         $content = $this->replaceVars(__DIR__.'/build/request.tpl');
         $content = str_replace('{REQUEST_RULE}', var_export($this->getRequestRule(), true), $content);
         $content = str_replace('{REQUEST_RULE_MESSAGE}', var_export($this->getRequestRuleMessage(), true), $content);
-        $file = $this->getVar('REQUEST_PATH').$this->model.'Request.php';
         file_put_contents($file, $content);
         $this->info('request create successflly');
     }
